Initial AutoMode CommandGroup Implementation
Handles driving forward while elevating, rotating then driving
forward to a set distance.

Was only tested on a 'bot on blocks, and only for setup
on the right side.

Further work to fully test for all start conditions on the
real practice field.
Also needed: implementation of grabber to fully complete the
automode logic.