Drive Elevator Up in Auto
To prevent sagging when getting to the middle circuit breaker,
rather than finishing the command (which would stop the motors & cause sagging)
continue to drive the motor at 40% up.

Since this is for auto mode, the time until the command is cancelled
will be on the order of 10seconds; so chances of burning out the motor
are limited.